% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_buffer.R
\name{add_buffer}
\alias{add_buffer}
\title{Add buffer to word}
\usage{
add_buffer(x, buffer = 5, fill = 0)
}
\arguments{
\item{x}{The word matrix}

\item{buffer}{The buffer ammount in pixles}

\item{fill}{The value to fill the buffer with}
}
\value{
A matrix with buffer applied
}
\description{
Add a simplified buffer to the extremites of a word matrix
}
