% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap_spiral_coordinates.R
\name{bitmap_spiral}
\alias{bitmap_spiral}
\title{Bitmap spiral coordinates}
\usage{
bitmap_spiral(x1, y1, step = 10, l = 100, type = sample(c(1, 2), 1))
}
\arguments{
\item{x1}{Starting x position}

\item{y1}{Starting y position}

\item{step}{Step length of spiral (distance before turning a right angle)}

\item{l}{Length of spiral (how many straight lines of length step) every 4 l gets a new layer of the spiral}

\item{type}{1 = square, 2 = diamond}
}
\description{
Generate a tibble of spiral cartesian coordinates as either a square spiral or a 
diamond spiral (a square spiral rotated by 45 degrees) All coordiabtes returned are
integers
}
\examples{
library(tidyverse)
# spiral_coords(0, 0, l=1, step = 1) \%>\%
#     mutate(xend = lead(x),
#            yend=lead(y)) \%>\%
#     ggplot(aes(x=x, y=y, xend=xend, yend=yend))+
#     geom_segment()+
#     geom_point(aes(x=0, y=0), col=2, size=2)+
#     coord_equal()
}
