% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap_wordcloud.R
\name{bmwc}
\alias{bmwc}
\title{Bitmap wordcloud}
\usage{
bmwc(data, min_size = 10, max_size = 200, angle_range = c(0, 0),
  buffer = 5, view = TRUE, save = FALSE, filename,
  font_face = "regular", font_family = "sans", seed = NULL,
  spiral_length = 50, spiral_step = 50, page_res = 600,
  v_buffer = 40, h_buffer = 30)
}
\arguments{
\item{data}{}

\item{min_size}{}

\item{max_size}{}

\item{angle_range}{}

\item{buffer}{}

\item{view}{}

\item{save}{}

\item{filename}{}

\item{font_face}{}

\item{font_family}{}

\item{seed}{}

\item{spiral_length}{}

\item{spiral_step}{}

\item{page_res}{}

\item{v_buffer}{}

\item{h_buffer}{}
}
\description{
Bitmap wordcloud
}
