% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_transform.R
\name{coord_trans}
\alias{coord_trans}
\title{Coordinate transform from word to page}
\usage{
coord_trans(px, py, word_m)
}
\arguments{
\item{px}{Page starting x-coordinate of word}

\item{py}{Page starting y-coordinate of word}

\item{word_m}{Word matrix}
}
\description{
Given a word matrix and its starting position on the page, this funtion
returns the rows and columns of the page matrix that will be occupied by the word matrix. The
page rowas and columns are calculated from placing the centre of the word matrix at the given
starting coordinates
}
