% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_share_folder.R
\name{create_and_share_folder}
\alias{create_and_share_folder}
\title{Create and Share Folder in Nextcloud}
\usage{
create_and_share_folder(
  webdav_url,
  folder_name,
  username,
  password,
  base_ocs_url
)
}
\arguments{
\item{webdav_url}{URL to the WebDAV endpoint.}

\item{folder_name}{The name of the folder to create.}

\item{username}{User's username.}

\item{password}{User's password.}

\item{base_ocs_url}{Base URL for the OCS API.}
}
\value{
A message indicating success or failure.
}
\description{
This function creates a new folder and shares it publicly using the Nextcloud WebDAV and OCS APIs.
}
