% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_word.R
\name{create_word}
\alias{create_word}
\title{Create a word from the all_glyphs data}
\usage{
create_word(word, weight)
}
\arguments{
\item{word}{word to create}

\item{weight}{pixel size}
}
\value{
a word matrix
}
\description{
Create a word from the all_glyphs data
}
\examples{
create_word("howdy", 50) \%>\% plot_mat()
}
