% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_extract_zip.R
\name{download_and_extract_zip}
\alias{download_and_extract_zip}
\title{Downloads and extracts external data to be used in the simulation}
\usage{
download_and_extract_zip(url, dest_folder = ".", zip_name = NULL)
}
\arguments{
\item{url}{The URL from where the external data can be downloaded.
The external data should be in a ZIP file.}

\item{dest_folder}{The folder where the data should be stored. Defaults to the current working directory.}

\item{zip_name}{Optional; specifies the name of the ZIP file if it differs from the last part of the URL.}
}
\value{
The path to the folder where the ZIP file was extracted, invisibly.
}
\description{
This function downloads a ZIP file from the specified URL and extracts its contents
into the designated folder. It is designed to handle data setup for simulations
by ensuring that the downloaded data is stored correctly.
}
\examples{
\dontrun{
download_and_extract_zip(url = "http://www.example.com/datafile.zip",
                         dest_folder = "path/to/destination")
}
}
