% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_collaborators.R
\name{invite_collaborators}
\alias{invite_collaborators}
\title{Invite collaborators to a GitHub repository}
\usage{
invite_collaborators(repo_name, collaborators, permission = "push")
}
\arguments{
\item{repo_name}{The name of the GitHub repository.}

\item{collaborators}{A character vector of GitHub usernames to be added as collaborators.}

\item{permission}{The permission level for the collaborators: 'pull', 'push', or 'admin'.}
}
\value{
Invisible NULL, primarily used for the side effects of adding collaborators.
}
\description{
This function invites one or more collaborators to a GitHub repository, specifying the level of access.
}
\examples{
invite_collaborators("mynewrepo", c("user1", "user2"), "push")
}
