% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_mat}
\alias{plot_mat}
\title{Image a matrix}
\usage{
plot_mat(m, ...)
}
\arguments{
\item{m}{The matrix to plot}

\item{...}{Further arguments passed to image()}
}
\description{
A wrapper around image() that plots a given matrix
}
