% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap_spiral_coordinates.R
\name{plot_spiral}
\alias{plot_spiral}
\title{Add a spiral to a plot}
\usage{
plot_spiral(spiral, ...)
}
\arguments{
\item{spiral}{as made by bitmap_spiral()}

\item{...}{Further arguments passed to graphics::segments()}
}
\description{
Add a spiral to a plot
}
