% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirals.R
\name{spiral}
\alias{spiral}
\title{Spiral x-y coordinates}
\usage{
spiral(x, y, n, res, a0, a1)
}
\arguments{
\item{x}{Centre of whole spiral (x)}

\item{y}{Centre of whole spiral (n)}

\item{n}{Number of full rotations of spiral}

\item{res}{Number of points to generate}

\item{a0}{Radius to start of spiral}

\item{a1}{Radius to end of spiral}
}
\description{
Parametereised spiral in cartesian coordinates
}
\examples{
sp <- spiral(x=0, y=0, n=5, res=500, a0=0.5, a1=3)
plot(sp$x, sp$y, type="l", asp=1)
}
