% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_zipped_directory.R
\name{upload_zipped_directory}
\alias{upload_zipped_directory}
\title{Upload a Zipped Directory to Nextcloud}
\usage{
upload_zipped_directory(
  local_folder_path,
  zip_file_path,
  nextcloud_base_url,
  target_nextcloud_folder,
  username
)
}
\arguments{
\item{local_folder_path}{The path to the local directory to be zipped.}

\item{zip_file_path}{Path where the ZIP file should be saved locally.}

\item{nextcloud_base_url}{The base URL of the Nextcloud instance.}

\item{target_nextcloud_folder}{The target folder in Nextcloud where the ZIP file will be uploaded.}

\item{username}{Nextcloud username.}
}
\value{
A message indicating the outcome of the upload operation.
}
\description{
This function zips a local directory and uploads the resulting ZIP file to a specified
location in Nextcloud using WebDAV.
}
\examples{
\dontrun{
upload_zipped_directory("path/to/local/folder", "path/to/zipfile.zip",
                                     "https://nextcloud.instance.com", "/path/in/nextcloud/",
                                     "username") }
}
