% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{wc}
\alias{wc}
\title{Produce wordcloud}
\usage{
wc(words, weights, word_buffer = 0, lower_cex = 0.5, upper_cex = 2,
  buffer = 0.2, lo_col = "black", hi_col = "black", spiral_n = 8,
  spiral_res = 500, spiral_a0 = 0, spiral_a1 = 0.4, seed = NULL,
  vert = TRUE, plot_spirals = FALSE, markup = FALSE)
}
\arguments{
\item{words}{Character vector of words to use in word cloud}

\item{weights}{The weigths associated with each word - will be normalised and passed as the text cex}

\item{word_buffer}{A buffer to be added to each word. Default is 0. Negative values will make words plot over eachother}

\item{lower_cex}{Lower cex for lowest word weight}

\item{upper_cex}{Upper cex for highest word weight}

\item{buffer}{Buffer for plot region x dimension (npc units)}

\item{lo_col}{Colour for lowest word weight (a continuos colour scale from lo to hi will be constructed)}

\item{hi_col}{Colour for highest word weight (a continuos colour scale from lo to hi will be constructed)}

\item{spiral_n}{Number of full rotations in spiral}

\item{spiral_res}{Resilution of spirals}

\item{spiral_a0}{Spiral start relative to x coordinate}

\item{spiral_a1}{Spiral end}

\item{seed}{Seed for reproducabile plots}

\item{vert}{Logical. Use vertical words in wordcloud}

\item{plot_spirals}{Logical. Choose whether to plott he spirals for each point on the page}

\item{markup}{Logical. Add word borders, x-y coordinates and page boundaries/buffer zone}
}
\description{
Produce wordcloud
}
\examples{
wc(obama$word[1:100],
obama$count[1:100],
lower_cex = 1,
upper_cex = 3.5,
spiral_n = 10,
spiral_res = 500,
lo_col = "grey75",
seed = 1)
}
