% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_matrix.R
\name{word_mat}
\alias{word_mat}
\title{Word matrix}
\usage{
word_mat(word, weight, rot, ...)
}
\arguments{
\item{word}{Character string}

\item{weight}{Word size to be passed to pixel_size}

\item{rot}{Word angle of rotation to be passed to rot}

\item{...}{Further parameters to be passed to fontr::glyph_bitmap()}
}
\description{
A wrapper around fontr::glyph_bitmap(). After the matrix is extracted, it is 'unclassed' (back to an S3 matrix)
and then 0s (the background to the character glyphs) are set as NA
}
