% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_upload_x0.R
\name{zip_upload_x0}
\alias{zip_upload_x0}
\title{Zip a Folder and Upload to a File Hosting Service}
\usage{
zip_upload_x0(folder, zip_file, upload_link = "https://x0.at/")
}
\arguments{
\item{folder}{Character. Path to the folder to be zipped and uploaded.}

\item{zip_file}{Character. Path (including filename) where the zip file should be saved.}

\item{upload_link}{Character. URL of the file hosting service (default: "https://x0.at/").}
}
\value{
Character. Shareable link to the uploaded zip file.
}
\description{
This function zips all files in a specified folder (including subfolders),
uploads the zip file to a specified file hosting service (default: x0.at),
and returns a shareable link.
}
\examples{
upload_x0("path/to/folder", "path/to/output.zip")
upload_x0("path/to/folder", "path/to/output.zip", "https://custom-upload-service.com/")
}
