% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap_wordcloud.R
\name{bmwc}
\alias{bmwc}
\title{Bitmap wordcloud}
\usage{
bmwc(data, min_size = 25, max_size = 300, angle_range = c(0, 0),
  buffer = 5, show_word_buffer = FALSE, view = TRUE, save = FALSE,
  filename, font_face = "regular", font_family = "sans", seed = NULL,
  v_buffer = 45, h_buffer = 35, page_w = 5000, page_h = 5000,
  spiral_radius = 1000, spiral_turns = 5, spiral_res = 500,
  markup = FALSE, plot_spirals = FALSE)
}
\arguments{
\item{data}{}

\item{min_size}{}

\item{max_size}{}

\item{angle_range}{}

\item{buffer}{}

\item{view}{}

\item{save}{}

\item{filename}{}

\item{font_face}{}

\item{font_family}{}

\item{seed}{}

\item{v_buffer}{}

\item{h_buffer}{}

\item{spiral_length}{}

\item{spiral_step}{}

\item{page_res}{}
}
\description{
Bitmap wordcloud
}
