% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createghrepo.R
\name{create_ghrepo}
\alias{create_ghrepo}
\title{Create a new GitHub repository and set an upstream}
\usage{
create_ghrepo(projname, upstream, collaborators = character())
}
\arguments{
\item{projname}{The name of the project. It need to be the same as in the upstream link.}

\item{upstream}{The URL of the upstream repository to set for pushing changes.}

\item{collaborators}{A character vector of GitHub usernames to be added as collaborators.}
}
\value{
Invisible NULL, used for side effects of creating and setting up the repository.
}
\description{
This function initializes a new GitHub repository with the specified name,
renames the local master branch to main, sets a remote named upstream,
and pushes the main branch to the newly set upstream.
}
\examples{
\dontrun{
create_ghrepo("mynewrepo", "https://github.com/username/mynewrepo.git")
}
}
